Move reusable bundle related classes in a dedicated package
